+++ src/objectbuilders/queryBuilder.ts
@@ -65,11 +65,11 @@ export class QueryBuilder extends ObjectBuilder {
 
         writer.writeStartNameSourceBlock("dataitem", dataItemName, writer.encodeName(tableSymbol.symbolName));
 
-        tableSymbol.childItems.forEach(
+        let fieldList : ALSymbolInfo[] = [];
+        tableSymbol.getAllSymbolsByKind(ALSymbolKind.Field, fieldList);
+        fieldList.forEach(
             item => {
-                if (item.alKind == ALSymbolKind.Field) {
-                    writer.writeNameSourceBlock("column", writer.createName(item.symbolName), writer.encodeName(item.symbolName));
-                }
+                writer.writeNameSourceBlock("column", writer.createName(item.symbolName), writer.encodeName(item.symbolName));
             }
         );